When fixing up a FDO class definition before applying to a fresh data store, we are applying a max ceiling on the given scale and precision if they are greater than the max allowed precision/scale in the provider's capabilities.
However if a provider reports a value of -1 for max precision or scale, that value would've been assigned straight the to precision or scale of the DataPropertyDefinition, which forbids negative values from being assigned.

To fix this, we will interpret -1 as being un-bounded, so we will only clamp the scale/precision value only if the max allowed scale/precision is > 0. Otherwise we'll leave the value as-is.

Fixes #84